Refactor module_list and core_module_list
