+++ tests/test_payment.py
@@ -309,7 +309,7 @@ async def test_webhook_payment_callback(
         payment_callback=callback,
     )
     mocker.patch(
-        "app.core.payment.endpoints_payment.module_list",
+        "app.core.payment.endpoints_payment.all_modules",
         [test_module],
     )
 
@@ -350,7 +350,7 @@ async def test_webhook_payment_callback_fail(
         payment_callback=callback,
     )
     mocker.patch(
-        "app.core.payment.endpoints_payment.module_list",
+        "app.core.payment.endpoints_payment.all_modules",
         [test_module],
     )
